+++ deployments/backend.yml
@@ -165,6 +165,17 @@ Resources:
     Properties:
       TemplateURL: web/ecr.yml
 
+  SourceApi:
+    Type: AWS::CloudFormation::Stack
+    Properties:
+      Parameters:
+        CloudWatchLogRetentionDays: !Ref CloudWatchLogRetentionDays
+        Debug: !Ref Debug
+        LayerVersionArns: !Join [',', !Ref LayerVersionArns]
+        TracingMode: !Ref TracingMode
+        SQSKeyId: !Ref QueueEncryptionKey
+      TemplateURL: core/source_api.yml
+
   AnalysisAPI:
     Type: AWS::CloudFormation::Stack
     Properties:
