+++ deployments/core/appsync.yml
@@ -127,15 +127,15 @@ Resources:
       LambdaConfig:
         LambdaFunctionArn: !Sub arn:${AWS::Partition}:lambda:${AWS::Region}:${AWS::AccountId}:function:panther-outputs-api
 
-  SnapshotAPILambdaDataSource:
+  SourceAPILambdaDataSource:
     Type: AWS::AppSync::DataSource
     Properties:
       ApiId: !GetAtt GraphQLApi.ApiId
-      Name: PantherSnapshotAPILambda
+      Name: PantherSourceAPILambda
       Type: AWS_LAMBDA
       ServiceRoleArn: !GetAtt AppsyncServiceRole.Arn
       LambdaConfig:
-        LambdaFunctionArn: !Sub arn:${AWS::Partition}:lambda:${AWS::Region}:${AWS::AccountId}:function:panther-snapshot-api
+        LambdaFunctionArn: !Sub arn:${AWS::Partition}:lambda:${AWS::Region}:${AWS::AccountId}:function:panther-source-api
 
   OrganizationAPILambdaDataSource:
     Type: AWS::AppSync::DataSource
@@ -460,7 +460,7 @@ Resources:
       ApiId: !GetAtt GraphQLApi.ApiId
       TypeName: Query
       FieldName: integrations
-      DataSourceName: !GetAtt SnapshotAPILambdaDataSource.Name
+      DataSourceName: !GetAtt SourceAPILambdaDataSource.Name
       RequestMappingTemplate: |
         {
           "version" : "2017-02-28",
@@ -485,7 +485,7 @@ Resources:
       ApiId: !GetAtt GraphQLApi.ApiId
       TypeName: Mutation
       FieldName: addIntegration
-      DataSourceName: !GetAtt SnapshotAPILambdaDataSource.Name
+      DataSourceName: !GetAtt SourceAPILambdaDataSource.Name
       RequestMappingTemplate: |
         #set ($sourcesInput = [])
         #foreach($item in $ctx.args.input.integrations)
@@ -521,7 +521,7 @@ Resources:
       ApiId: !GetAtt GraphQLApi.ApiId
       TypeName: Mutation
       FieldName: updateIntegration
-      DataSourceName: !GetAtt SnapshotAPILambdaDataSource.Name
+      DataSourceName: !GetAtt SourceAPILambdaDataSource.Name
       RequestMappingTemplate: |
         #set ($input = $util.defaultIfNull($ctx.args.input, {}))
         $util.qr($input.put("scanIntervalMins", 1440))
@@ -547,7 +547,7 @@ Resources:
       ApiId: !GetAtt GraphQLApi.ApiId
       TypeName: Mutation
       FieldName: deleteIntegration
-      DataSourceName: !GetAtt SnapshotAPILambdaDataSource.Name
+      DataSourceName: !GetAtt SourceAPILambdaDataSource.Name
       RequestMappingTemplate: |
         {
           "version" : "2017-02-28",
