+++ deployments/core/source_api.yml
@@ -1,0 +1,167 @@
+# Panther is a scalable, powerful, cloud-native SIEM written in Golang/React.
+# Copyright (C) 2020 Panther Labs Inc
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+AWSTemplateFormatVersion: 2010-09-09
+Transform: AWS::Serverless-2016-10-31
+Description: Source API
+
+Parameters:
+  CloudWatchLogRetentionDays:
+    Type: Number
+    Description: CloudWatch log retention period
+    Default: 365
+  Debug:
+    Type: String
+    Default: false
+    Description: Toggle debug logging
+    AllowedValues: [true, false]
+  LayerVersionArns:
+    Type: CommaDelimitedList
+    Description: List of LayerVersion ARNs to attach to each function
+    Default: ''
+  TracingMode:
+    Type: String
+    Description: Enable XRay tracing on Lambda and API Gateway
+    AllowedValues: ['', Active, PassThrough]
+    Default: ''
+  SQSKeyId:
+    Type: String
+    Description: KMS key ID for SQS encryption
+
+  AuditRoleName:
+    Type: String
+    Default: PantherAuditRole
+    Description: The name of the IAM role to assume to perform AWS account/service scans
+  RemediationRoleName:
+    Type: String
+    Default: PantherRemediationRole
+    Description: The name of the IAM role to assume to perform AWS account/service scans
+  CWERoleName:
+    Type: String
+    Default: PantherCloudFormationStackSetExecutionRole
+    Description: The name of the IAM role to assume to perform AWS account/service scans
+  LogProcessingRoleName:
+    Type: String
+    Default: PantherLogProcessingRole
+    Description: The name of the IAM role to assume to perform AWS account/service scans
+  LogAnalysisQueueName:
+    Type: String
+    Default: panther-input-data-notifications-queue
+  SnapshotPollerQueueName:
+    Type: String
+    Default: panther-snapshot-queue
+
+Conditions:
+  AttachLayers: !Not [!Equals [!Join ['', !Ref LayerVersionArns], '']]
+  TracingEnabled: !Not [!Equals ['', !Ref TracingMode]]
+
+Resources:
+  ##### Source API #####
+  IntegrationsTable:
+    Type: AWS::DynamoDB::Table
+    Properties:
+      TableName: panther-source-integrations
+      # <cfndoc>
+      # This table does hold the configured accounts and log sources for monitoring.
+      #
+      # Failure Impact
+      # * Processing of policies could be slowed or stopped if there are errors/throttles.
+      # * The Panther user interface could be impacted.
+      # </cfndoc>
+      BillingMode: PAY_PER_REQUEST
+      AttributeDefinitions:
+        - AttributeName: integrationId
+          AttributeType: S
+      KeySchema:
+        - AttributeName: integrationId
+          KeyType: HASH
+      PointInTimeRecoverySpecification:
+        PointInTimeRecoveryEnabled: True
+      SSESpecification: # Enable server-side encryption
+        SSEEnabled: True
+
+  ApiLambdaFunction:
+    Type: AWS::Serverless::Function
+    Properties:
+      CodeUri: ../../out/bin/internal/core/source_api/main
+      Description: Manages database of source integrations
+      Environment:
+        Variables:
+          DEBUG: !Ref Debug
+          SNAPSHOT_POLLERS_QUEUE_URL: !Sub https://sqs.${AWS::Region}.amazonaws.com/${AWS::AccountId}/${SnapshotPollerQueueName}
+          LOG_PROCESSOR_QUEUE_URL: !Sub https://sqs.${AWS::Region}.amazonaws.com/${AWS::AccountId}/${LogAnalysisQueueName}
+          LOG_PROCESSOR_QUEUE_ARN: !Sub arn:${AWS::Partition}:sqs:${AWS::Region}:${AWS::AccountId}:${LogAnalysisQueueName}
+          TABLE_NAME: !Ref IntegrationsTable
+      FunctionName: panther-source-api
+      # <cfndoc>
+      # The `panther-source-api` lambda manages Cloud Security and Log Analysis sources. This includes
+      # creating, testing, updating, listing, and deleting sources.
+      #
+      # Failure Impact
+      # * Failure of this lambda will prevent sources from being manageable, and will interrupt daily scans.
+      # </cfndoc>
+      Handler: main
+      Layers: !If [AttachLayers, !Ref LayerVersionArns, !Ref 'AWS::NoValue']
+      MemorySize: 512
+      Runtime: go1.x
+      Timeout: 120
+      Tracing: !If [TracingEnabled, !Ref TracingMode, !Ref 'AWS::NoValue']
+      Policies:
+        - Id: IntegrationsTablePermissions
+          Version: 2012-10-17
+          Statement:
+            - Effect: Allow
+              Action:
+                - dynamodb:*Item
+                - dynamodb:Query
+                - dynamodb:Scan
+              Resource: !GetAtt IntegrationsTable.Arn
+        - Id: SendSQSMessages
+          Version: 2012-10-17
+          Statement:
+            - Effect: Allow
+              Action:
+                - sqs:SendMessage
+                - sqs:SendMessageBatch
+              Resource: !Sub arn:${AWS::Partition}:sqs:${AWS::Region}:${AWS::AccountId}:${SnapshotPollerQueueName}
+            - Effect: Allow
+              Action:
+                - kms:Decrypt
+                - kms:GenerateDataKey
+              Resource: !Sub arn:${AWS::Partition}:kms:${AWS::Region}:${AWS::AccountId}:key/${SQSKeyId}
+        - Id: UpdateLogProcessorQueue
+          Version: 2012-10-17
+          Statement:
+            - Effect: Allow
+              Action: sqs:*QueueAttributes
+              Resource: !Sub arn:${AWS::Partition}:sqs:${AWS::Region}:${AWS::AccountId}:${LogAnalysisQueueName}
+        - Id: AssumePantherAuditRoles
+          Version: 2012-10-17
+          Statement:
+            - Effect: Allow
+              Action: sts:AssumeRole
+              Resource:
+                - !Sub arn:${AWS::Partition}:iam::*:role/${AuditRoleName}
+                - !Sub arn:${AWS::Partition}:iam::*:role/${RemediationRoleName}
+                - !Sub arn:${AWS::Partition}:iam::*:role/${CWERoleName}
+                - !Sub arn:${AWS::Partition}:iam::*:role/${LogProcessingRoleName}
+
+  ApiLogGroup:
+    Type: AWS::Logs::LogGroup
+    Properties:
+      LogGroupName: /aws/lambda/panther-source-api
+      RetentionInDays: !Ref CloudWatchLogRetentionDays
+
