+++ internal/compliance/aws_event_processor/processor/accounts.go
@@ -28,14 +28,14 @@ import (
 	"github.com/aws/aws-sdk-go/service/s3"
 	"go.uber.org/zap"
 
-	"github.com/panther-labs/panther/api/lambda/snapshot/models"
+	"github.com/panther-labs/panther/api/lambda/source/models"
 	"github.com/panther-labs/panther/pkg/genericapi"
 )
 
 const (
 	cweAccountTimeout       = 15 * time.Minute
 	refreshInterval         = 2 * time.Minute
-	snapshotAPIFunctionName = "panther-snapshot-api"
+	snapshotAPIFunctionName = "panther-source-api"
 )
 
 var (
