+++ internal/compliance/aws_event_processor/processor/accounts_test.go
@@ -30,7 +30,7 @@ import (
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 
-	"github.com/panther-labs/panther/api/lambda/snapshot/models"
+	"github.com/panther-labs/panther/api/lambda/source/models"
 )
 
 var (
@@ -164,7 +164,7 @@ func getTestInvokeInput() *lambda.InvokeInput {
 	}
 
 	return &lambda.InvokeInput{
-		FunctionName: aws.String("panther-snapshot-api"),
+		FunctionName: aws.String("panther-source-api"),
 		Payload:      payload,
 	}
 }
