+++ internal/core/source_api/api/list_integrations_test.go
@@ -28,9 +28,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/panther-labs/panther/api/lambda/snapshot/models"
-	"github.com/panther-labs/panther/internal/compliance/snapshot_api/ddb"
-	"github.com/panther-labs/panther/internal/compliance/snapshot_api/ddb/modelstest"
+	"github.com/panther-labs/panther/api/lambda/source/models"
+	"github.com/panther-labs/panther/internal/core/source_api/ddb"
+	"github.com/panther-labs/panther/internal/core/source_api/ddb/modelstest"
 )
 
 func TestListIntegrations(t *testing.T) {
