+++ internal/core/source_api/api/sqs_utils.go
