+++ internal/core/source_api/api/update_integration_test.go
@@ -29,9 +29,9 @@ import (
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 
-	"github.com/panther-labs/panther/api/lambda/snapshot/models"
-	"github.com/panther-labs/panther/internal/compliance/snapshot_api/ddb"
-	"github.com/panther-labs/panther/internal/compliance/snapshot_api/ddb/modelstest"
+	"github.com/panther-labs/panther/api/lambda/source/models"
+	"github.com/panther-labs/panther/internal/core/source_api/ddb"
+	"github.com/panther-labs/panther/internal/core/source_api/ddb/modelstest"
 )
 
 func TestUpdateIntegrationSettings(t *testing.T) {
