+++ tools/dashboards/infra.go
@@ -61,7 +61,7 @@ var infraJSON = `
             "height": 3,
             "properties": {
                 "metrics": [
-                    [ "AWS/Lambda", "Invocations", "FunctionName", "panther-snapshot-api", "Resource", "panther-snapshot-api", { "stat": "Sum", "region": "us-east-1" } ]
+                    [ "AWS/Lambda", "Invocations", "FunctionName", "panther-source-api", "Resource", "panther-source-api", { "stat": "Sum", "region": "us-east-1" } ]
                 ],
                 "region": "us-east-1",
                 "title": "Invocations",
@@ -77,9 +77,9 @@ var infraJSON = `
             "height": 3,
             "properties": {
                 "metrics": [
-                    [ "AWS/Lambda", "Duration", "FunctionName", "panther-snapshot-api", "Resource", "panther-snapshot-api", { "stat": "Minimum", "region": "us-east-1" } ],
-                    [ "AWS/Lambda", "Duration", "FunctionName", "panther-snapshot-api", "Resource", "panther-snapshot-api", { "stat": "Average", "region": "us-east-1" } ],
-                    [ "AWS/Lambda", "Duration", "FunctionName", "panther-snapshot-api", "Resource", "panther-snapshot-api", { "stat": "Maximum", "region": "us-east-1" } ]
+                    [ "AWS/Lambda", "Duration", "FunctionName", "panther-source-api", "Resource", "panther-source-api", { "stat": "Minimum", "region": "us-east-1" } ],
+                    [ "AWS/Lambda", "Duration", "FunctionName", "panther-source-api", "Resource", "panther-source-api", { "stat": "Average", "region": "us-east-1" } ],
+                    [ "AWS/Lambda", "Duration", "FunctionName", "panther-source-api", "Resource", "panther-source-api", { "stat": "Maximum", "region": "us-east-1" } ]
                 ],
                 "region": "us-east-1",
                 "view": "timeSeries",
@@ -95,8 +95,8 @@ var infraJSON = `
             "height": 3,
             "properties": {
                 "metrics": [
-                    [ "AWS/Lambda", "Errors", "FunctionName", "panther-snapshot-api", "Resource", "panther-snapshot-api", { "id": "errors", "stat": "Sum", "color": "#d13212", "region": "us-east-1" } ],
-                    [ "AWS/Lambda", "Invocations", "FunctionName", "panther-snapshot-api", "Resource", "panther-snapshot-api", { "id": "invocations", "stat": "Sum", "visible": false, "region": "us-east-1" } ],
+                    [ "AWS/Lambda", "Errors", "FunctionName", "panther-source-api", "Resource", "panther-source-api", { "id": "errors", "stat": "Sum", "color": "#d13212", "region": "us-east-1" } ],
+                    [ "AWS/Lambda", "Invocations", "FunctionName", "panther-source-api", "Resource", "panther-source-api", { "id": "invocations", "stat": "Sum", "visible": false, "region": "us-east-1" } ],
                     [ { "expression": "100 - 100 * errors / MAX([errors, invocations])", "label": "Success rate (%)", "id": "availability", "yAxis": "right", "region": "us-east-1" } ]
                 ],
                 "region": "us-east-1",
@@ -288,7 +288,7 @@ var infraJSON = `
             "width": 3,
             "height": 3,
             "properties": {
-                "query": "SOURCE '/aws/lambda/panther-snapshot-api' | filter component like 'snapshot' | stats max(percentMemUsed) as used by bin(5min)\n",
+                "query": "SOURCE '/aws/lambda/panther-source-api' | filter component like 'snapshot' | stats max(percentMemUsed) as used by bin(5min)\n",
                 "region": "us-east-1",
                 "title": "Memory Usage (%)",
                 "view": "timeSeries",
@@ -302,7 +302,7 @@ var infraJSON = `
             "width": 3,
             "height": 3,
             "properties": {
-                "query": "SOURCE '/aws/lambda/panther-snapshot-api' | filter component like 'snapshot' | stats max(heapSizeMB) as heap by bin(5min)\n",
+                "query": "SOURCE '/aws/lambda/panther-source-api' | filter component like 'snapshot' | stats max(heapSizeMB) as heap by bin(5min)\n",
                 "region": "us-east-1",
                 "title": "Heap Usage (MB)",
                 "view": "timeSeries",
@@ -453,7 +453,7 @@ var infraJSON = `
             "width": 9,
             "height": 6,
             "properties": {
-                "query": "SOURCE '/aws/lambda/panther-snapshot-api' | SOURCE '/aws/lambda/panther-snapshot-pollers' | SOURCE '/aws/lambda/panther-snapshot-scheduler' | SOURCE '/aws/lambda/panther-aws-event-processor' | SOURCE '/aws/lambda/panther-resources-api' | SOURCE '/aws/lambda/panther-resource-processor' | SOURCE '/aws/lambda/panther-policy-engine' | filter  @message like '[ERROR]' or  @message like '[WARN]' or level='error' or level='warn'\n| fields @timestamp, @message\n| sort @timestamp desc",
+                "query": "SOURCE '/aws/lambda/panther-source-api' | SOURCE '/aws/lambda/panther-snapshot-pollers' | SOURCE '/aws/lambda/panther-snapshot-scheduler' | SOURCE '/aws/lambda/panther-aws-event-processor' | SOURCE '/aws/lambda/panther-resources-api' | SOURCE '/aws/lambda/panther-resource-processor' | SOURCE '/aws/lambda/panther-policy-engine' | filter  @message like '[ERROR]' or  @message like '[WARN]' or level='error' or level='warn'\n| fields @timestamp, @message\n| sort @timestamp desc",
                 "region": "us-east-1",
                 "stacked": false,
                 "title": "Most Recent 20 Errors and Warnings",
@@ -467,7 +467,7 @@ var infraJSON = `
             "width": 9,
             "height": 6,
             "properties": {
-                "query": "SOURCE '/aws/lambda/panther-aws-event-processor' | SOURCE '/aws/lambda/panther-snapshot-api' | SOURCE '/aws/lambda/panther-snapshot-pollers' | SOURCE '/aws/lambda/panther-snapshot-scheduler' | SOURCE '/aws/lambda/panther-resources-api' | SOURCE '/aws/lambda/panther-resource-processor' | SOURCE '/aws/lambda/panther-policy-engine' | filter  @message like '[ERROR]' or  @message like '[WARN]' or level='error'  or level='warn'\n| stats sum(strcontains(level, 'error')+strcontains(@message, '[ERROR]')) as errors, sum(strcontains(level, 'warn')+strcontains(@message, '[WARN]')) as warns by bin(5m)",
+                "query": "SOURCE '/aws/lambda/panther-aws-event-processor' | SOURCE '/aws/lambda/panther-source-api' | SOURCE '/aws/lambda/panther-snapshot-pollers' | SOURCE '/aws/lambda/panther-snapshot-scheduler' | SOURCE '/aws/lambda/panther-resources-api' | SOURCE '/aws/lambda/panther-resource-processor' | SOURCE '/aws/lambda/panther-policy-engine' | filter  @message like '[ERROR]' or  @message like '[WARN]' or level='error'  or level='warn'\n| stats sum(strcontains(level, 'error')+strcontains(@message, '[ERROR]')) as errors, sum(strcontains(level, 'warn')+strcontains(@message, '[WARN]')) as warns by bin(5m)",
                 "region": "us-east-1",
                 "stacked": false,
                 "title": "Errors and Warnings",
